+++ CHANGES.md
@@ -1,10 +1,11 @@
 # 1.1.0
 
 -   Switch to github container registry
--   Use Node 12
+-   Use Node 16
 -   Multi-arch docker image release
 -   #16 strip html tags from license field
 -   trim publisher name to avoid unnecessary blank
+-   #1 make sure generated date string that doesn't match JSON schema (ISO 8601 format)
 
 # 1.0.0